# openapi.yaml
openapi: 3.0.3
info:
  title: tr-engine API
  version: 0.9.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  description: |
    Backend service for trunk-recorder data ingestion and querying.
    Provides REST APIs for accessing radio system data, calls, talkgroups,
    units, and recorder state.

    ## Authentication

    All endpoints (except `/health` and `/auth-init`) require a bearer token
    via the `Authorization: Bearer <token>` header or `?token=` query param.

    **Two-tier auth (optional):** When `WRITE_TOKEN` is configured, mutating
    operations (POST, PATCH, PUT, DELETE) require the write token. Read
    operations (GET) accept either the read token (`AUTH_TOKEN`) or write
    token. If `WRITE_TOKEN` is not set, all operations accept `AUTH_TOKEN`
    (backwards compatible).

    The `/auth-init` endpoint returns only the read token — the write token
    is never exposed by any endpoint.

    ## Data Model: Systems and Sites

    The data model separates **logical radio networks** from **recording
    sites**:

    - **System** (`system_id`): a logical radio network. For P25, this
      is the WACN-level network identified by `(sysid, wacn)`. For
      conventional, it is identified by `(instance_id, sys_name)`.
      Talkgroups and units belong to the system level because they are
      shared across all sites on the same P25 network.

    - **Site** (`site_id`): a recording point within a system. Each
...
      specific site.

...
    ```
...

...

...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /calls:
    get:
      operationId: listCalls
      summary: List calls
      description: |
        Returns call recordings with comprehensive filters. This is the
        primary endpoint for querying call history.

        **Common patterns:**
        - Live feed: `?sort=-stop_time&deduplicate=true&limit=50`
        - Talkgroup history: `?tgid=9178&sysid=348&start_time=...&end_time=...`
        - Unit activity: `?unit_id=924003`
        - Emergency calls: `?emergency=true&sort=-start_time`
      tags: [calls]
      parameters:
        - name: sysid
          in: query
          description: Filter by P25 SYSID (comma-separated for multiple, e.g., "348,34D"). Alias "sysids" also accepted.
          schema:
            type: string
        - name: system_id
          in: query
          description: Filter by system database ID (comma-separated for multiple). Alias "systems" also accepted.
          schema:
            type: string
            example: "1,2"
        - name: site_id
          in: query
          description: Filter by recording site ID (comma-separated for multiple). Alias "sites" also accepted.
          schema:
            type: string
            example: "1,3"
        - name: tgid
          in: query
          description: Filter by talkgroup ID (comma-separated for multiple). Alias "tgids" also accepted.
          schema:
            type: string
            example: "9178,5344"
        - name: unit_id
          in: query
          description: Filter by unit radio ID (comma-separated for multiple). Aliases "units", "unit_ids" also accepted.
          schema:
            type: string
            example: "924003,104"
        - name: emergency
          in: query
          description: Filter by emergency flag
          schema:
            type: boolean
        - name: encrypted
          in: query
          description: Filter by encryption status
          schema:
            type: boolean
        - name: deduplicate
          in: query
          description: |
            When true, returns one call per call_group (best recording).
            Useful for feed views where simulcast duplicates are unwanted.
          schema:
            type: boolean
            default: false
        - $ref: "#/components/parameters/startTime"
        - $ref: "#/components/parameters/endTime"
        - name: sort
          in: query
          description: |
            Sort field. Prefix with `-` for descending.
            Allowed: start_time, stop_time, duration, tgid, freq
          schema:
            type: string
            default: "-start_time"
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/offset"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CallListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /calls/active:
    get:
      operationId: listActiveCalls
      summary: List active calls (real-time)
      description: |
        Returns currently active calls from the in-memory MQTT tracker.
        This is a live endpoint — it returns a point-in-time snapshot of
        all active calls from memory. There is no pagination; the full
        list is returned in a single response.

        Active calls may have incomplete fields (no stop_time, no audio_url,
        duration is elapsed-so-far). The response uses the same Call schema
        with nullable fields where data is not yet available.
      tags: [calls]
      parameters:
        - name: sysid
          in: query
          description: Filter by P25 SYSID
          schema:
            type: string
        - name: tgid
          in: query
          description: Filter by talkgroup ID
          schema:
            type: integer
        - name: emergency
          in: query
          description: Filter by emergency flag
          schema:
            type: boolean
        - name: encrypted
          in: query
          description: Filter by encryption status
          schema:
            type: boolean
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActiveCallListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /calls/{id}:
    get:
      operationId: getCall
      summary: Get a call
      description: |
        Returns a single call recording by its database ID.
      tags: [calls]
      parameters:
        - $ref: "#/components/parameters/callId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Call"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /calls/{id}/audio:
    get:
      operationId: getCallAudio
      summary: Stream call audio
      description: Streams the audio file for a call. Content-Type varies by source format.
      tags: [calls]
      parameters:
        - $ref: "#/components/parameters/callId"
      responses:
        "200":
          description: Audio stream
          content:
            audio/mp4:
              schema:
                type: string
                format: binary
            audio/mpeg:
              schema:
                type: string
                format: binary
            audio/wav:
              schema:
                type: string
                format: binary
            audio/ogg:
              schema:
                type: string
                format: binary
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /calls/{id}/frequencies:
    get:
      operationId: getCallFrequencies
      summary: Get call frequency list
      description: |
        Returns the frequency entries (freqList) within a call, ordered by
        position in the audio.
      tags: [calls]
      parameters:
        - $ref: "#/components/parameters/callId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CallFrequencyListResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /calls/{id}/transmissions:
    get:
      operationId: getCallTransmissions
      summary: Get call transmission list
      description: |
        Returns the unit transmissions (srcList) within a call, ordered by
        position in the audio. Each entry represents a unit keying up.
        Includes unit alpha_tags.
      tags: [calls]
      parameters:
        - $ref: "#/components/parameters/callId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CallTransmissionListResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

  # ----------------------------------------------------------
  # Transcriptions
  # ----------------------------------------------------------
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        Bearer token authentication. Pass token in the Authorization header:
        `Authorization: Bearer <token>`

  # ----------------------------------------------------------
  # Reusable Parameters
  # ----------------------------------------------------------
  parameters:
    callId:
      name: id
      in: path
      required: true
      description: Call record database ID
      schema:
        type: integer
        example: 48531
...
    startTime:
      name: start_time
      in: query
      description: Start of time range (RFC 3339)
      schema:
        type: string
        format: date-time
    endTime:
      name: end_time
      in: query
      description: End of time range (RFC 3339)
      schema:
        type: string
        format: date-time
    limit:
      name: limit
      in: query
      description: Results per page
      schema:
        type: integer
        default: 50
        minimum: 1
        maximum: 1000
    offset:
      name: offset
      in: query
      description: Page offset (number of results to skip)
      schema:
        type: integer
        default: 0
        minimum: 0
...
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    InternalError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"

  # ----------------------------------------------------------
  # Schemas
  # ----------------------------------------------------------
  schemas:
...
    CallState:
      type: string
      enum: [monitoring, recording, stopped, completed]
      description: |
        Call recording state (mapped from trunk-recorder integer values):
        - **monitoring**: control channel activity seen, no audio capture yet
        - **recording**: actively capturing audio
        - **stopped**: recording ended, finalizing
        - **completed**: audio file written, ready for access
    MonState:
      type: string
      enum: [unspecified, active, duplicate]
      description: |
        Call monitor state (from trunk-recorder):
        - **unspecified**: default / not specifically monitored
        - **active**: actively monitored
        - **duplicate**: duplicate call on another recorder
...
    Error:
      type: object
      required: [code, error]
      properties:
        code:
          type: string
          description: Machine-readable error code for programmatic handling
          example: not_found
          enum:
            - bad_request
            - unauthorized
            - forbidden
            - not_found
            - conflict
            - rate_limited
            - internal_error
            - service_unavailable
            - invalid_body
            - invalid_parameter
            - invalid_time_range
            - query_failed
            - ambiguous_id
            - duplicate
            - request_timeout
        error:
          type: string
          example: Resource not found
        detail:
          type: string
          description: Additional context when available
...
    Call:
      type: object
      description: A recorded radio call/transmission
      required: [call_id, system_id, tgid, start_time]
      properties:
        # Identifiers
        call_id:
          type: integer
          description: Unique database ID for this recording
          example: 48531
        call_group_id:
          type: integer
          description: |
            Call group this recording belongs to. Multiple recordings of the
            same transmission (from different sites) share a call_group_id.
            Derived from (system_id, tgid, start_unix).
          example: 1

        # System context (embedded to avoid lookup)
        system_id:
          type: integer
          description: Logical radio system this call belongs to
          example: 1
        system_name:
          type: string
          description: System display name
          example: "Butler/Warren P25"
        sysid:
          type: string
          description: P25 SYSID (for display; "0" for conventional)
          example: "348"

        # Site context (which recorder captured this)
        site_id:
          type: integer
          description: Recording site that captured this call
          example: 1
        site_short_name:
          type: string
          description: TR sys_name of the capturing site (e.g., "butco", "warco")
          example: butco

        # Talkgroup context (embedded to avoid lookup)
        tgid:
          type: integer
          example: 9178
        tg_alpha_tag:
          type: string
          description: Talkgroup short-form display name
          example: "09-8L Main"
        tg_description:
          type: string
          description: Human-readable talkgroup description
          example: Middletown Police Dispatch
        tg_tag:
          type: string
          description: Functional category tag (e.g., Law Dispatch, Fire Tac)
          example: Law Dispatch
        tg_group:
          type: string
          description: Broad organizational grouping
          example: "Butler County (09) Law"

        # Timing
        start_time:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
        stop_time:
          type: string
          format: date-time
          nullable: true
          description: "Null for active (in-progress) calls"
          example: "2024-01-15T10:30:45Z"
        duration:
          type: number
          description: Duration in seconds (elapsed so far for active calls)
          example: 45.5

        # Audio
        audio_url:
          type: string
          nullable: true
          description: "Relative URL to audio endpoint. Null for active calls."
          example: /api/v1/calls/48531/audio
        audio_type:
          type: string
          description: "Audio encoding type from trunk-recorder (e.g., `digital`, `digital tdma`, `analog`)"
          example: digital
        audio_size:
          type: integer
          description: Audio file size in bytes
          example: 45000

        # Signal quality
        freq:
          type: integer
          description: Frequency in Hz
          example: 851012500
        freq_error:
          type: integer
          description: Frequency error in Hz
          example: 0
        signal_db:
          type: number
          nullable: true
          description: Signal strength in dB (null if unavailable; MQTT sends 999 as sentinel)
          example: -23
        noise_db:
          type: number
          nullable: true
          description: Noise floor in dB (null if unavailable; MQTT sends 999 as sentinel)
          example: -50
        error_count:
          type: integer
          description: Decode error count
          example: 2
        spike_count:
          type: integer
          description: Audio spike count
          example: 0

        # Status
        call_state:
          $ref: "#/components/schemas/CallState"
        mon_state:
          $ref: "#/components/schemas/MonState"
        emergency:
          type: boolean
          example: false
        encrypted:
          type: boolean
          example: false
        analog:
          type: boolean
          example: false
        conventional:
          type: boolean
          example: false
        phase2_tdma:
          type: boolean
          example: false
        tdma_slot:
          type: integer
          description: "TDMA slot (0=slot 1 or N/A, 1=slot 2)"
          example: 0

        # Patches
        patched_tgids:
          type: array
          description: Other talkgroup IDs patched into this call
          items:
            type: integer

        # Transmission and frequency data (JSONB columns)
        src_list:
          type: array
          description: |
            Source/transmission list from trunk-recorder's srcList.
            Each entry is a unit transmission segment within the call.
            Also available via GET /calls/{id}/transmissions.
          items:
            $ref: "#/components/schemas/CallTransmission"
        freq_list:
          type: array
          description: |
            Frequency list from trunk-recorder's freqList.
            Each entry is a frequency segment within the call.
            Also available via GET /calls/{id}/frequencies.
          items:
            $ref: "#/components/schemas/CallFrequency"
        unit_ids:
          type: array
          description: Denormalized array of distinct unit radio IDs (src values) from src_list. GIN-indexed for fast unit filtering.
          items:
            type: integer

        # Units on this call (embedded)
        units:
          type: array
          items:
            $ref: "#/components/schemas/CallUnit"

        # Transcription (denormalized from transcriptions table)
        has_transcription:
          type: boolean
          description: Whether this call has been transcribed
          example: true
        transcription_status:
          $ref: "#/components/schemas/TranscriptionStatus"
        transcription_text:
          type: string
          nullable: true
          description: Primary transcription text (null if no transcription)
          example: Engine 5 responding to the scene
        transcription_word_count:
          type: integer
          nullable: true
          example: 6

        # Extensible metadata
        metadata_json:
          type: object
          additionalProperties: true
          description: Arbitrary metadata from trunk-recorder
    CallUnit:
      type: object
      description: A unit that transmitted during a call
      required: [system_id, unit_id]
      properties:
        system_id:
          type: integer
          description: Database system ID for disambiguation in multi-system deployments
          example: 1
        unit_id:
          type: integer
          description: Unit radio ID (RID)
          example: 924003
        alpha_tag:
          type: string
          description: Unit display name
          example: "09 7COM3"
...
    CallFrequency:
      type: object
      description: |
        A frequency segment from trunk-recorder's freqList. Each entry
        represents a period of time the call spent on a given frequency.
      properties:
        freq:
          type: integer
          description: Frequency in Hz
          example: 154875000
        time:
          type: string
          format: date-time
          description: When this frequency segment started (RFC 3339)
          example: "2024-04-15T19:03:22Z"
        pos:
          type: number
          description: Position in the audio file in seconds
          example: 0.0
        len:
          type: number
          description: Duration on this frequency in seconds
          example: 3.24
        error_count:
          type: integer
          example: 50
        spike_count:
          type: integer
          example: 3
    CallTransmission:
      type: object
      description: |
        A unit key-up from trunk-recorder's srcList. Each entry represents
        one unit's transmission within the call. The `src` field maps to
        the unit radio ID (RID).
      properties:
        src:
          type: integer
          description: Unit radio ID (RID) — matches unit_id in other models
          example: 104
        tag:
          type: string
          description: Unit alpha tag at time of transmission (from trunk-recorder)
          example: "09 7COM3"
        time:
          type: string
          format: date-time
          description: When this unit keyed up (RFC 3339)
          example: "2024-04-15T19:03:22Z"
        pos:
          type: number
          description: Position in audio file in seconds
          example: 0.0
        duration:
          type: number
          description: Duration of this transmission segment in seconds (computed from gap to next transmission or end of call)
          example: 3.5
        emergency:
          type: integer
          description: "Emergency flag (0=normal, 1=emergency)"
          example: 0
        signal_system:
          type: string
          description: Signal system identifier (typically empty)
          example: ""
...
    TranscriptionStatus:
      type: string
      enum: [none, auto, reviewed, verified, excluded]
      description: |
        - **none**: no transcription attempted
        - **auto**: machine-transcribed, unreviewed
        - **reviewed**: passed quality checks
        - **verified**: human-verified correct
        - **excluded**: permanently excluded from dataset
...
    CallListResponse:
      type: object
      required: [calls, total, limit, offset]
      properties:
        calls:
          type: array
          items:
            $ref: "#/components/schemas/Call"
        total:
          type: integer
          description: Total matching results (across all pages)
          example: 2500
        limit:
          type: integer
          example: 50
        offset:
          type: integer
          example: 0
    ActiveCallListResponse:
      type: object
      required: [calls, total]
      properties:
        calls:
          type: array
          description: |
            Active calls from MQTT tracker. Fields like stop_time and
            audio_url will be null since the call is still in progress.
          items:
            $ref: "#/components/schemas/Call"
        total:
          type: integer
          description: Number of currently active calls
          example: 5
...
    CallFrequencyListResponse:
      type: object
      required: [frequencies, total]
      properties:
        frequencies:
          type: array
          items:
            $ref: "#/components/schemas/CallFrequency"
        total:
          type: integer
          example: 3
    CallTransmissionListResponse:
      type: object
      required: [transmissions, total]
      properties:
        transmissions:
          type: array
          items:
            $ref: "#/components/schemas/CallTransmission"
        total:
          type: integer
          example: 4

    # --- Stats responses ---
...